week_1/requirements.txt
google-generativeai
python-dotenv
transformers
nltk
streamlit

